--- src/js/styles/components/shepherd-cancel-link.js
@@ -1,10 +1,0 @@
-export function shepherdCancelLinkCSS() {
-  return {
-    fontSize: '2em',
-    textDecoration: 'none',
-    transition: 'color 0.5s ease',
-    '&::before': {
-      content: '"\u00d7"'
-    }
-  };
-}
