--- src/js/styles/components/shepherd-footer.js
@@ -1,14 +1,0 @@
-export function shepherdFooterCSS(variables) {
-  return {
-    borderBottomLeftRadius: variables.shepherdElementBorderRadius,
-    borderBottomRightRadius: variables.shepherdElementBorderRadius,
-    display: 'flex',
-    justifyContent: 'flex-end',
-    padding: '0 0.75em 0.75em',
-    '.shepherd-button': {
-      '&:last-child': {
-        marginRight: 0
-      }
-    }
-  };
-}
